requirements.txt
flask==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Security-Too==5.3.3
flask-admin==1.6.1